requirements.txt
facebook-sdk==3.1.0
Flask==2.1.0
openai==0.10.2
python-dotenv